requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
pytest==7.4.3
py==1.11.0
pytest-selenium==4.0.1
urllib3==2.0.7
click==8.1.7